+++ projects/ng-charts/src/lib/components/pieChart.ts
@@ -1,15 +1,17 @@
-import { Component, OnInit, Input } from '@angular/core'
+import { Component, OnInit, Input, AfterViewInit } from '@angular/core'
 declare const d3
 declare const $
+declare const tippy
 
 @Component({
   selector: 'lib-pie-chart',
   template: `<svg />`,
   styles: ['svg { width: 100%; height: 100%; }']
 })
-export class NgPieChartComponent implements OnInit {
+export class NgPieChartComponent implements OnInit, AfterViewInit {
   d3
   $
+  tippy
   @Input() data
   @Input() title
 
@@ -46,6 +48,12 @@ export class NgPieChartComponent implements OnInit {
     arc.append('path')
       .attr('d', path)
       .attr('fill', (d) => color(d.data.percentage))
+      .attr('title', (d) => `
+        <div style="text-align: left;">
+          fruit: <b>${d.data.fruit}</b><br>
+          percentage: <b>${d.data.percentage}</b>
+        </div>
+      `)
       .style('opacity', '0.8')
       .on('mouseover mousemove', function(d) {
         $(this).css('opacity', '1.0')
@@ -63,5 +71,15 @@ export class NgPieChartComponent implements OnInit {
     }
 
     return svg.node()
+  }
+
+  ngAfterViewInit() {
+    tippy('[title]', {
+      content(reference) {
+        const title = reference.getAttribute('title');
+        reference.removeAttribute('title');
+        return title;
+      },
+    });
   }
 }