+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add tooltips for Bar-Chart
 - Add tooltips for Line-Chart
 - Add tooltips for Heatmap
+- Add tooltips for Pie-Chart
 
 ## [0.2.3] - 2019-10-17